+++ mrp_sale_info/models/__init__.py
@@ -4,3 +4,4 @@
 
 from . import mrp_production
 from . import mrp_workorder
+from . import procurement
