+++ mrp_sale_info/models/mrp_production.py
@@ -10,32 +10,10 @@ class MrpProduction(models.Model):
 
     sale_id = fields.Many2one(
         'sale.order',
-        compute='_compute_sale_info',
         string='Sale order',
-        readonly=True,
-        store=True)
+        readonly=True)
     partner_id = fields.Many2one(
         'res.partner',
-        compute='_compute_sale_info',
-        string='Customer',
-        store=True)
-    commitment_date = fields.Datetime(compute='_compute_sale_info',
-                                      string='Commitment Date')
-
-    @api.multi
-    def _compute_sale_info(self):
-        def get_parent_move(move):
-            if move.move_dest_id:
-                return get_parent_move(move.move_dest_id)
-            return move
-
-        for production in self:
-            move = get_parent_move(production.move_finished_ids)
-            proc = move.procurement_id
-            production.sale_id = proc and proc.sale_line_id and \
-                proc.sale_line_id.order_id.id or False
-            production.partner_id = production.sale_id and \
-                production.sale_id.partner_id and \
-                production.sale_id.partner_id.id or False
-            production.commitment_date = production.sale_id and \
-                production.sale_id.commitment_date or ''
+        readonly=True,
+        string='Customer')
+    commitment_date = fields.Datetime(string='Commitment Date')
